+++ src/main/java/com/wenxin2/marioverse/entities/EepCheepEntity.java
@@ -167,7 +167,7 @@ public class EepCheepEntity extends CheepCheepEntity implements GeoEntity {
         return this.entityData.get(SIZE);
     }
 
-    public void setSize(String variant) {
-        this.entityData.set(SIZE, variant);
+    public void setSize(String size) {
+        this.entityData.set(SIZE, size);
     }
 }