+++ changelog.md
@@ -2,6 +2,10 @@
 - Migrated compat to use Curios instead of Accessories
 - Added the data component `marioverse:undying_charm`
   - Any item with this component will function as a one up
+- Added Piranha Plant variants
+  - Tropical variants spawn in jungles & lush caves
+  - Cave variants spawn underground
+  - Deep Cave variants spawn deep underground
 
 ## Update 1.11.0
 - Added power up spawn eggs